Add solutions for problems 332 to 356

+++ cp/problem_333.cpp
@@ -33,7 +33,7 @@ using pll = pair<ll,ll>;
 // ------------------------- SOLVE -------------------------- 
 
 void solve(){
-    int n,k,x;
+    ll n,k,x;
     get(n);
     get(k);
     get(x);
